requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
pandas>=1.3.0
gdown
dvc
mlflow
notebook
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
scipy
Flask
Flask-Cors
scikit-learn==1.3.0
seaborn==0.12.2